requirements.txt
customtkinter
pyvista
scipy
pandas
reportlab
trimesh
rtree
pyembree
easydict
pyyaml
xxhash